PROJECT/requirements.txt
streamlit>=1.25.0
spacy>=3.6.0
nltk>=3.8.0
scikit-learn>=1.3.0
pdfplumber>=0.8.0
python-docx>=0.8.12
textblob>=0.17.1
fpdf>=1.7.2
